+++ Day16 - The Floor Will Be Lava/Puzzle/PointExtensions.cs
@@ -1,4 +1,4 @@
-namespace AdventOfCode.Year2023.Day16;
+namespace AdventOfCode.Year2023.Day16.Puzzle;
 
 internal static class PointExtensions
 {
